+++ src/renderer/mixins/index.js
@@ -1,4 +1,5 @@
 import base from './mixin-base';
 import resource from './mixin-resource';
+import contextmenu from './mixin-contextmenu';
 
-export {base, resource};
+export {base, resource, contextmenu};
